backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-docx==1.1.0
PyMuPDF==1.24.2
transformers==4.41.2
sentence-transformers==2.7.0
faiss-cpu==1.7.4
tqdm==4.66.4
pydantic==1.10.13
python-multipart==0.0.9
langchain==0.2.1
langchain-community==0.2.1
